requirements.txt
python-dotenv
spotdl
spotipy
py-sonic
owncloud